WIP: Separate services from pipelines
Developing and deploying cd pipelines is a different type of activity
from developing and deploying the service stacks. In addition to this,
we are now creating separate foundation stacks for test and prod, but not
specifically instantiating these stacks in the cdk app is an anti-pattern.
Putting all of these things in base.ts was becoming harder to parse and
understand. This is an attempt at separating these types of activities,
without completely separating pipelines into their own app (which is an
option but would require bigger changes). The idea is that you will now
have to pass an additional context variable called `stackType`. The
available stacks to deploy will depend on what is given for this context
variable, either `service` or `pipeline`. `service` will be the default
to retain backward compatibility with both devs and pipelines.

Here's the output of `cdk list` as an example:
```
$ cdk list -c env=dev
marble-elastic
marble-foundation
marble-image-processing
marble-image-service
marble-manifest
marble-redbox
marble-user-content
marble-website

$ cdk list -c env=dev -c stackType=pipeline
marble-image-processing-deployment
marble-image-service-deployment
marble-prod-foundation
marble-test-foundation
marble-user-content-deployment
marble-redbox-deployment
marble-website-deployment
```
